update the blink_esp32 sketch nd the DelayState library source
I have modified the DelayState library in DelayState.h and
DelayState.cpp in folder Lib source so that the DelayClass is now using
the StateClass for all state related functionality. The idea is to
separate out the Delay functionality which may or may not use the State
functionality and the State functionality which can now be used without
the Delay functionality.

Now using three LEDs for blinking at different rates, two on a
breadboard and the internal LED on pin 2.

Have also added code to monitor a momentary contact press button which
will detect the first button press then count button presses for five
seconds.

There is a problem with the current button press functionaity in that we
are sometimes seeing multiple button press indications when there was
only one press. We need to added debounce functionality to eliminate
spurious button press indications.